Refactor sandbox controller into logical files for clarity
Split the monolithic sandbox_controller.go (1266 lines) into:
- sandbox_constants.go: Condition types, reasons, labels, finalizer
- sandbox_pod.go: Pod creation and sidecar injection
- sandbox_status.go: Status patching and condition determination
- sandbox_shutdown.go: Finalization, shutdown policy, timeouts
- sandbox_controller.go: Core reconciliation, template, network policy
